Message appears when trying to execute/export program with syntax errors

+++ lib/eco/debug.py
@@ -13,22 +13,23 @@ class Debugger(QObject):
         # incoming signals
         self.signal_command = SIGNAL("command")
         self.signal_start = SIGNAL("start")
-        # outcoming signals
+        # outgoing signals
         self.signal_done = SIGNAL("finished")
         self.signal_output = SIGNAL("output")
         self.signal_toggle_buttons = SIGNAL("disablebuttons")
+        self.signal_execute_fail = SIGNAL("executefail")
         self.telnet = None
 
     def start_pdb(self):   
         self.breakpoints = {'keep': [], 'del': []}      
         self.line_read_until = '(Pdb)'   
         self.process = self.window.getEditor().tm.export(debug=True)
-        flags = fcntl(self.process.stdout, F_GETFL)
-        fcntl(self.process.stdout, F_SETFL, flags | O_NONBLOCK)       
         if not self.process:
-            self.emit(self.signal_output, "Cannot run debugger process, cannot debug.")
+            self.emit(self.signal_execute_fail)
             self.exit()
             return
+        flags = fcntl(self.process.stdout, F_GETFL)
+        fcntl(self.process.stdout, F_SETFL, flags | O_NONBLOCK)
         self.telnet = None
         attempt = 0
         while not self.telnet:
@@ -85,8 +86,9 @@ class Debugger(QObject):
 
     def print_program_output(self):
         try:
-            program_output = self.process.stdout.read()
-            self.emit(self.signal_output, program_output.strip())
+            if self.process:
+                program_output = self.process.stdout.read()
+                self.emit(self.signal_output, program_output.strip())
         except (OSError, IOError):
             pass
         return
@@ -156,7 +158,8 @@ class Debugger(QObject):
     def exit(self):
         self.print_program_output()
         try:
-            self.process.kill()
+            if self.process:
+                self.process.kill()
         except OSError:
             pass
         if self.window.debugging:
